rename to ArtworkRail
